requirements.txt
torch
torchvision
pillow
opencv-python
pyyaml
streamlit
numpy
